+++ sql/schema/004_order_items.sql
@@ -4,9 +4,9 @@ CREATE TABLE order_items (
     order_id INTEGER NOT NULL REFERENCES orders(id) ON DELETE CASCADE,
     item TEXT NOT NULL,
     quantity INTEGER NOT NULL DEFAULT 1,
-    price DECIMAL NOT NULL,
-    discount DECIMAL NOT NULL DEFAULT 0.0, -- Percentage discount on the item
-    item_total DECIMAL NOT NULL,
+    price FLOAT NOT NULL,
+    discount FLOAT NOT NULL DEFAULT 0.0, -- Percentage discount on the item
+    item_total FLOAT NOT NULL,
     pocket_number INTEGER -- If item is on the customer location's planogram, this is the pocket number
 );
